Basic ship and player controllers
Got the basic ship and player controllers in place for the prototype. Not polished at all, ship moves forward, and can rotate left and right. The player is a first person controller, where the camera is moved round with the mouse. The player can control the ship by going to the helm and pressing F, their control will then switch to that of the ships and they won't be able to walk around until they press F again. One large problem is that when the ship rotates, due to the solution to keeping the player on the ship being that it is a child of the ship, the player's rotation looks slightly weird.