Large Player/Ship Controller Changes
Tested the parent/child rotation and postion sharing between the ship and the player, ended up just adding on the rotation around the y axis of the ship onto the player's rotation in mouseLook(), check to see if they are a child of the ship before doing so. Cleaned up and commented the code, and made it so that when the player takes control of the ship, they are turned to face the front of it. Finally, added in checks to only allow one player to control the ship at any given time.